+++ apps/opentelemetry/src/otel_configuration.erl
@@ -51,6 +51,9 @@
                views := list(), %% TODO: type should be `[otel_meter_server:view_config]'
                                 %% when Metrics are moved out of the experimental app
                readers := [#{id := atom(), module => module(), config => map()}],
+               exemplars_enabled := boolean(),
+               exemplar_filter := always_on | always_off | trace_based,
+
                processors := list(),
                sampler := {atom(), term()},
                sweeper := #{interval => integer() | infinity,
@@ -89,6 +92,8 @@ new() ->
                    metrics_exporter => {opentelemetry_exporter, #{}},
                    views => [],
                    readers => [],
+                   exemplars_enabled => false,
+                   exemplar_filter => trace_based,
                    processors => [{otel_batch_processor, ?BATCH_PROCESSOR_DEFAULTS}],
                    sampler => {parent_based, #{root => always_on}},
                    sweeper => #{interval => timer:minutes(10),
@@ -311,6 +316,8 @@ config_mappings(general_sdk) ->
      {"OTEL_METRICS_EXPORTER", metrics_exporter, exporter},
      {"OTEL_METRIC_VIEWS", views, views},
      {"OTEL_METRIC_READERS", readers, readers},
+     {"OTEL_ERLANG_X_EXEMPLARS_ENABLED", exemplars_enabled, boolean},
+     {"OTEL_METRICS_EXEMPLAR_FILTER", exemplar_filter, exemplar_filter},
      {"OTEL_RESOURCE_DETECTORS", resource_detectors, existing_atom_list},
      {"OTEL_RESOURCE_DETECTOR_TIMEOUT", resource_detector_timeout, integer},
 
@@ -499,6 +506,14 @@ transform(span_processor, SpanProcessor) ->
     SpanProcessor;
 transform(readers, Readers) ->
     Readers;
+transform(exemplar_filter, ExemplarFilter) when ExemplarFilter =:= "always_on" ;
+                                                ExemplarFilter =:= always_on ->
+    always_on;
+transform(exemplar_filter, ExemplarFilter) when ExemplarFilter =:= "always_off" ;
+                                                ExemplarFilter =:= always_off ->
+    always_off;
+transform(exemplar_filter, _) ->
+    trace_based;
 transform(views, Views) ->
     Views.