+++ apps/opentelemetry_api_experimental/src/otel_counter.erl
@@ -19,9 +19,9 @@
 -module(otel_counter).
 
 -export([create/3,
-         add/2,
          add/3,
-         add/4]).
+         add/4,
+         add/5]).
 
 -include("otel_metrics.hrl").
 -include_lib("kernel/include/logger.hrl").
@@ -33,20 +33,20 @@
 create(Meter, Name, Opts) ->
     otel_meter:create_counter(Meter, Name, Opts).
 
--spec add(otel_instrument:t(), pos_integer() | float()) -> ok.
-add(Instrument=#instrument{module=Module}, Number) ->
-    Module:record(Instrument, Number).
+-spec add(otel_ctx:t(), otel_instrument:t(), pos_integer() | float()) -> ok.
+add(Ctx, Instrument=#instrument{module=Module}, Number) ->
+    Module:record(Ctx, Instrument, Number).
 
 -spec add(
-    otel_meter:t() | otel_instrument:t(),
+    otel_ctx:t(), otel_meter:t() | otel_instrument:t(),
     otel_instrument:name() | pos_integer() | float(),
     pos_integer() | float() | opentelemetry:attributes_map()) -> ok.
-add(Instrument=#instrument{module=Module}, Number, Attributes) ->
-    Module:record(Instrument, Number, Attributes);
+add(Ctx, Instrument=#instrument{module=Module}, Number, Attributes) ->
+    Module:record(Ctx, Instrument, Number, Attributes);
 
-add(Meter, Name, Number) ->
-    otel_meter:record(Meter, Name, Number).
+add(Ctx, Meter, Name, Number) ->
+    otel_meter:record(Ctx, Meter, Name, Number).
 
--spec add(otel_meter:t(), otel_instrument:name(), pos_integer() |float(), opentelemetry:attributes_map()) -> ok.
-add(Meter, Name, Number, Attributes) ->
-    otel_meter:record(Meter, Name, Number, Attributes).
+-spec add(otel_ctx:t(), otel_meter:t(), otel_instrument:name(), pos_integer() |float(), opentelemetry:attributes_map()) -> ok.
+add(Ctx, Meter, Name, Number, Attributes) ->
+    otel_meter:record(Ctx, Meter, Name, Number, Attributes).