+++ apps/opentelemetry_experimental/src/otel_aggregation_histogram_explicit.erl
@@ -20,8 +20,9 @@
 -behaviour(otel_aggregation).
 
 -export([init/2,
-         aggregate/4,
-         collect/3]).
+         aggregate/7,
+         collect/4,
+         default_buckets/0]).
 
 -include("otel_metrics.hrl").
 -include_lib("opentelemetry_api_experimental/include/otel_metrics.hrl").
@@ -133,14 +134,17 @@
 
 %% ignore eqwalizer errors in functions using a lot of matchspecs
 -eqwalizer({nowarn_function, checkpoint/3}).
--eqwalizer({nowarn_function, collect/3}).
+-eqwalizer({nowarn_function, collect/4}).
 -dialyzer({nowarn_function, checkpoint/3}).
--dialyzer({nowarn_function, aggregate/4}).
--dialyzer({nowarn_function, collect/3}).
+-dialyzer({nowarn_function, aggregate/7}).
+-dialyzer({nowarn_function, collect/4}).
 -dialyzer({nowarn_function, maybe_delete_old_generation/4}).
--dialyzer({nowarn_function, datapoint/2}).
+-dialyzer({nowarn_function, datapoint/4}).
 -dialyzer({nowarn_function, get_buckets/2}).
 -dialyzer({nowarn_function, counters_get/2}).
+
+default_buckets() ->
+    ?DEFAULT_BOUNDARIES.
 
 init(#stream{name=Name,
              reader=ReaderId,
@@ -166,10 +170,11 @@ init(#stream{name=Name,
                                     sum=0
                                    }.
 
-aggregate(Table, #stream{name=Name,
-                         reader=ReaderId,
-                         aggregation_options=Options,
-                         forget=Forget}, Value, Attributes) ->
+aggregate(Ctx, Table, ExemplarsTab, #stream{name=Name,
+                                            reader=ReaderId,
+                                            aggregation_options=Options,
+                                            forget=Forget,
+                                            exemplar_reservoir=ExemplarReservoir}, Value, Attributes, DroppedAttributes) ->
     Generation = case Forget of
                      true ->
                          otel_metric_reader:checkpoint_generation(ReaderId);
@@ -195,7 +200,13 @@ aggregate(Table, #stream{name=Name,
             counters:add(BucketCounts, BucketIdx, 1),
 
             MS = ?AGGREATE_MATCH_SPEC(Key, Value, BucketCounts),
-            1 =:= ets:select_replace(Table, MS)
+            case ets:select_replace(Table, MS) of
+                1 ->
+                    otel_metric_exemplar_reservoir:offer(Ctx, ExemplarReservoir, ExemplarsTab, Key, Value, DroppedAttributes),
+                    true;
+                _ ->
+                    false
+            end
     end.
 
 checkpoint(Tab, #stream{name=Name,
@@ -235,10 +246,11 @@ checkpoint(_Tab, _, _) ->
 
     ok.
 
-collect(Tab, Stream=#stream{name=Name,
-                            reader=ReaderId,
-                            temporality=Temporality,
-                            forget=Forget}, Generation0) ->
+collect(Tab, ExemplarsTab, Stream=#stream{name=Name,
+                                          reader=ReaderId,
+                                          temporality=Temporality,
+                                          forget=Forget,
+                                          exemplar_reservoir=ExemplarReservoir}, Generation0) ->
     CollectionStartTime = opentelemetry:timestamp(),
     Generation = case Forget of
                      true ->
@@ -259,7 +271,7 @@ collect(Tab, Stream=#stream{name=Name,
                                                max='$8',
                                                sum='$9'}, [], ['$_']}],
     AttributesAggregation = ets:select(Tab, Select),
-    Result = #histogram{datapoints=[datapoint(CollectionStartTime, SumAgg) || SumAgg <- AttributesAggregation],
+    Result = #histogram{datapoints=[datapoint(ExemplarReservoir, ExemplarsTab, CollectionStartTime, SumAgg) || SumAgg <- AttributesAggregation],
                         aggregation_temporality=Temporality},
 
     %% would be nice to do this in the reader so its not duplicated in each aggregator
@@ -282,16 +294,18 @@ maybe_delete_old_generation(Tab, Name, ReaderId, Generation) ->
                [true]}],
     ets:select_delete(Tab, Select).
 
-datapoint(CollectionStartTime, #explicit_histogram_aggregation{
-                                  key={_, Attributes, _, _},
-                                  explicit_bucket_boundaries=Boundaries,
-                                  start_time=StartTime,
-                                  checkpoint=undefined,
-                                  bucket_counts=BucketCounts,
-                                  min=Min,
-                                  max=Max,
-                                  sum=Sum
-                                 }) ->
+datapoint(ExemplarReservoir, ExemplarsTab, CollectionStartTime,
+          #explicit_histogram_aggregation{
+             key=Key={_, Attributes, _, _},
+             explicit_bucket_boundaries=Boundaries,
+             start_time=StartTime,
+             checkpoint=undefined,
+             bucket_counts=BucketCounts,
+             min=Min,
+             max=Max,
+             sum=Sum
+            }) ->
+    Exemplars = otel_metric_exemplar_reservoir:collect(ExemplarReservoir, ExemplarsTab, Key),
     Buckets = get_buckets(BucketCounts, Boundaries),
     #histogram_datapoint{
        attributes=Attributes,
@@ -301,20 +315,22 @@ datapoint(CollectionStartTime, #explicit_histogram_aggregation{
        sum=Sum,
        bucket_counts=Buckets,
        explicit_bounds=Boundaries,
-       exemplars=[],
+       exemplars=Exemplars,
        flags=0,
        min=Min,
        max=Max
       };
-datapoint(CollectionStartTime, #explicit_histogram_aggregation{
-                                  key={_, Attributes, _, _},
-                                  explicit_bucket_boundaries=Boundaries,
-                                  checkpoint=#explicit_histogram_checkpoint{bucket_counts=BucketCounts,
-                                                                            min=Min,
-                                                                            max=Max,
-                                                                            sum=Sum,
-                                                                            start_time=StartTime}
-                                 }) ->
+datapoint(ExemplarReservoir, ExemplarsTab, CollectionStartTime,
+          #explicit_histogram_aggregation{
+             key=Key={_, Attributes, _, _},
+             explicit_bucket_boundaries=Boundaries,
+             checkpoint=#explicit_histogram_checkpoint{bucket_counts=BucketCounts,
+                                                       min=Min,
+                                                       max=Max,
+                                                       sum=Sum,
+                                                       start_time=StartTime}
+            }) ->
+    Exemplars = otel_metric_exemplar_reservoir:collect(ExemplarReservoir, ExemplarsTab, Key),
     Buckets = get_buckets(BucketCounts, Boundaries),
     #histogram_datapoint{
        attributes=Attributes,
@@ -324,7 +340,7 @@ datapoint(CollectionStartTime, #explicit_histogram_aggregation{
        sum=Sum,
        bucket_counts=Buckets,
        explicit_bounds=Boundaries,
-       exemplars=[],
+       exemplars=Exemplars,
        flags=0,
        min=Min,
        max=Max