+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Handle `explict_bucket_boundaries` advisory parameter](https://github.com/open-telemetry/opentelemetry-erlang/pull/628)
 - [Rename `boundaries` to `explict_bucket_boundaries` in histogram explicit aggregation options](https://github.com/open-telemetry/opentelemetry-erlang/pull/628)
 - [Allow creating wildcard views](https://github.com/open-telemetry/opentelemetry-erlang/pull/624)
+- [Exemplars support](https://github.com/open-telemetry/opentelemetry-erlang/pull/692)
 
 ### Changes
 
@@ -57,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - [Correctly record histogram values greater than last boundary](https://github.com/open-telemetry/opentelemetry-erlang/pull/614)
  - [Readers should use a default cumulative temporality if not specified](https://github.com/open-telemetry/opentelemetry-erlang/pull/613)
  - [Check for positive data values in counters and histograms](https://github.com/open-telemetry/opentelemetry-erlang/pull/632)
+- [Fix Delta metric export to only include those recorded in collection cycle](https://github.com/open-telemetry/opentelemetry-erlang/pull/677)
 
 ## SDK 1.3.1 - 2023-08-15